#64: First try to decouple the project into smaller subprojects. NOTE: Code does not compile! Just want to backup changes.
