+++ core/src/main/scala/org/zalando/jsonapi/json/package.scala
