+++ core/src/main/scala/org/zalando/jsonapi/model/package.scala
