+++ core/src/main/scala/org/zalando/jsonapi/package.scala
