+++ core/src/test/scala/org/zalando/jsonapi/json/Person.scala
