+++ spray-json/src/test/scala/org/zalando/jsonapi/sprayjson/SprayJsonJsonapiSupportSpec.scala
@@ -1,0 +1,11 @@
+package org.zalando.jsonapi.sprayjson
+
+import org.zalando.jsonapi.json.JsonapiSupportSpec
+import spray.json.CompactPrinter
+
+class SprayJsonJsonapiSupportSpec extends JsonapiSupportSpec {
+  implicit val printer = CompactPrinter
+  override implicit def jsonapiRootObjectMarshaller = SprayJsonJsonapiSupport.sprayJsonJsonapiMarshaller
+  override implicit def jsonapiRootObjectUnmarshaller = SprayJsonJsonapiSupport.sprayJsonJsonapiUnmarshaller
+  override def jsonapiSupportClassName = "SprayJsonJsonapiSupport"
+}
