--- src/main/scala/org/zalando/jsonapi/json/FieldNames.scala
@@ -1,28 +1,0 @@
-package org.zalando.jsonapi.json
-
-private[json] object FieldNames {
-  val `id` = "id"
-  val `type` = "type"
-  val `attributes` = "attributes"
-  val `links` = "links"
-  val `meta` = "meta"
-  val `relationships` = "relationships"
-  val `self` = "self"
-  val `about` = "about"
-  val `first` = "first"
-  val `next` = "next"
-  val `prev` = "prev"
-  val `last` = "last"
-  val `related` = "related"
-  val `status` = "status"
-  val `code` = "code"
-  val `title` = "title"
-  val `detail` = "detail"
-  val `source` = "source"
-  val `data` = "data"
-  val `errors` = "errors"
-  val `included` = "included"
-  val `jsonapi` = "jsonapi"
-  val `pointer` = "pointer"
-  val `parameter` = "parameter"
-}
